Revert "currently able to switch scenes. using "cheat" by pressing M"
This reverts commit 38a0346250d0641219164d79bf0c6dbceb5fc8b5.

+++ Game2/Src/Game2/CollisionSystem.cpp
@@ -572,15 +572,15 @@ namespace Engine
 									aabb2 = collisionComponent2->aabb;
 									if (entity2->GetID() == 7)
 									{
-										tower2CollidingEntityHealth = static_cast<int>(statsComponent2->health);
+										tower2CollidingEntityHealth = statsComponent2->health;
 									}
 									if (entity2->GetID() == 8)
 									{
-										tower1CollidingEntityHealth = static_cast<int>(statsComponent2->health);
+										tower1CollidingEntityHealth = statsComponent2->health;
 									}
 									if (entity2->GetID() == 9)
 									{
-										castleCollidingEntityHealth = static_cast<int>(statsComponent2->health);
+										castleCollidingEntityHealth = statsComponent2->health;
 									}
 								}
