+++ Game2/Src/Game2/PathfindingSystem.cpp
@@ -413,11 +413,11 @@ namespace Engine
                         
                         if (entity->GetID() == 7)
                         {
-                            tower2CollidingEntityHealth = static_cast<int>(statsComponent->health);
+                            tower2CollidingEntityHealth = statsComponent->health;
                         }
                         if (entity->GetID() == 8)
                         {
-                            tower1CollidingEntityHealth = static_cast<int>(statsComponent->health);
+                            tower1CollidingEntityHealth = statsComponent->health;
                             // std::cout << "Tower 2 health: " << tower2CollidingEntityHealth << std::endl;
                         }
